requirement.txt
python-dotenv
scrapy
bs4
pandas
pymongo
sentence_transformers
streamlit
openai
flask
flask-cors
spacy 
scikit-learn
pyvi
